integrate Ned's latest processing updates
* All of the hard coded analysis and plotting variables have been
moved to the top level wrapper function and are saved in a `par`
parameters structure.

* The code has been modified to write the plot data to a CSV file.
This file can be imported into a spreadsheet if desired.  It might be
good to add the software version to this file.

* I have added two path directories for the saving of (1) intermediate
analysis files and (2) the plot and CSV file.  These directories are
included in the par structure in the wrapper file.  For now, the default
paths are `pwd`.  The code still looks in the current directory to find
the raw data, though it would also be easy to change this location.

* The maximum frequency included in the plots and CSV output is now
determined automatically as the frequency when the phase velocity 95% CI
exceeds the 30% of average range.

* The last function that makes the plot and saves the text CSV file
has been renamed as `MakePlotAndSaveOutputCSVfile.m`.

* The last function also generates an output `errorFlag` which is
returned as `0` if the results are judged to be valid, and `1` if the
results are not valid.  Currently, the results are reported as not valid
if (1) the gSWS for either displacement or velocity are outside the
range 0.5 - 5 m/s (these limits are set in the `par` structure in the
wrapper) or (2) if the gSWS 95% CI exceeds the 30% of average for either
the displacement or velocity results.